refactor(spf): move text-track-segment-loader factory to media/actors
Phase 2 of decoupling text-track actors from DOM. The factory's runtime
had exactly one DOM dependency — the `parseVttSegment` import — and the
rest of its logic was already host-agnostic. Inject the parser instead
of importing it directly, and move the implementation to media/actors/.

Changes:

- media/actors/text-track-segment-loader.ts: now carries the factory
  implementation, generic over C extends Cue. Signature adds a
  `parseSegment: (url: string) => Promise<C[]>` parameter (typed as the
  new CueParser<C>). All core / media deps; no DOM imports.

- dom/actors/text-track-segment-loader.ts: becomes a thin 25-line
  wrapper that binds the browser's parseVttSegment into the media-level
  factory. Re-exports the types so existing call sites don't change.

- Error message: generalized from "Failed to load VTT segment" to
  "Failed to load text-track segment" since the injected parser need
  not be a VTT parser. Two test assertions updated to match.

Behavior is unchanged from a caller's perspective. Non-DOM engines
(future DASH, MoQ) can build the actor directly against the
media-level factory with their own cue parser — VTT isn't required
structurally.

Next (Phase 3): move load-text-track-cues behavior to media/behaviors/
with its DOM deps (parseSegment, actor factories, textTrack iteration)
declared via composition config. That's the last coupling between this
feature and the dom/ tree.

Build (161 files), tests (707), typecheck, workspace check, lint — all green.
